server/requirements.txt
flask_sqlalchemy==3.0.5
flask_wtf==1.1.1
flask==2.3.3
mysql-connector-python==8.0.28
python_dotenv==1.0.0
werkzeug==2.3.7